refactored toggle
