+++ src/components/Toggle/index.test.tsx
@@ -1,36 +1,34 @@
-import {fireEvent, render, screen} from '@testing-library/react';
+import { render, screen, fireEvent } from "@testing-library/react";
+import Toggle from ".";
 
-import Toggle from '.';
-
-describe('/src/components/Toggle', () => {
-  it('should properly render the component', () => {
+describe.only("should render Toggle component", () => {
+  it("renders toggle switch component", () => {
     render(<Toggle on={true} />);
-    const toggle = screen.getByTestId('toggle');
-    expect(toggle).toBeInTheDocument();
+    const labelElement = screen.getByTestId("toggle");
+    expect(labelElement).toBeInTheDocument();
   });
 
-  it('should properly display enabled status', () => {
+  it("display enabled status", () => {
     render(<Toggle on={true} />);
-    const toggle = screen.getByTestId('toggle') as HTMLInputElement;
-    expect(toggle.checked).toBe(true);
+    const labelElement = screen.getByTestId("toggle") as HTMLInputElement;
+    expect(labelElement.checked).toBe(true);
   });
-  
-  it('should properly display disabled status', () => {
+
+  it("display disabled status", () => {
     render(<Toggle on={false} />);
-    const toggle = screen.getByTestId('toggle') as HTMLInputElement;
-    expect(toggle.checked).toBe(false);
+    const labelElement = screen.getByTestId("toggle") as HTMLInputElement;
+    expect(labelElement.checked).toBe(false);
   });
-  
-  it('should properly trigger on state change event', () => {
+
+  it("should trigger state change", () => {
     const callback = jest.fn();
     render(<Toggle on={true} onStateChange={callback} />);
-    
-    const toggle = screen.getByTestId('toggle');
-    fireEvent.click(toggle);
-    fireEvent.click(toggle);
-    
+    const labelElement = screen.getByTestId("toggle");
+    fireEvent.click(labelElement);
+    fireEvent.click(labelElement);
+
     expect(callback).toHaveBeenCalledTimes(2);
     expect(callback).toHaveBeenNthCalledWith(1, false);
     expect(callback).toHaveBeenNthCalledWith(2, true);
   });
-})
+});
