+++ src/pages/TextSettings/index.tsx
@@ -5,7 +5,7 @@ import DropDownFontText from "../../components/InnerTextSettingPage/DropDownFont
 import COLORS from "../../constants/colors";
 import SIZE from "../../constants/size";
 import ColorPicker from "../../components/ColorPicker";
-import ToggleSwitch from "../../components/ToggleSwitch";
+import Toggle from "../../components/Toggle";
 
 const StyledTextSettingsContainer = styled.div`
   display: flex;
@@ -21,7 +21,7 @@ const StyledTextSettingsContainer = styled.div`
 export default function TextSettings() {
   return (
     <StyledTextSettingsContainer data-testid="textSettingsContainer">
-      <ToggleSwitch on={true} />
+      <Toggle on={true} />
       <p>The quick brown fox jumps over the lazy dog</p>
       <DropDownFontText />
       <ChangeSize />
